added usage example to graphics folder
