+++ public/usage-examples/graphics/get_font_style-1-example.txt
