+++ public/usage-examples/graphics/get_font_style-2-example.txt
